Move deletion functions to dedicated view files
